requirements.txt
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
dataclasses==0.6
Django==2.1.9
django-cors-headers==3.0.2
django-filter==2.1.0
django-pyodbc-azure==2.1.0.0
django-rest-framework==0.1.0
djangorestframework==3.9.4
elasticsearch==7.0.2
grpcio==1.21.1
gunicorn==19.9.0
httplib2==0.13.0
idna==2.8
msgpack==0.6.1
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyodbc==4.0.26
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.2.4
textblob==0.15.3
uritemplate==3.0.0
urllib3==1.25.3
